feat: endpoint extract phone user
