+++ src/main/java/com/micorservice/users/application/dto/response/PhoneResponseDto.java
@@ -1,0 +1,6 @@
+package com.micorservice.users.application.dto.response;
+
+public record PhoneResponseDto(
+        String phoneNumber
+) {
+}
