+++ src/main/java/com/micorservice/users/application/dto/response/RestaurantIdResponseDto.java
@@ -1,4 +1,4 @@
-package com.micorservice.users.application.dto.request;
+package com.micorservice.users.application.dto.response;
 
 public record RestaurantIdResponseDto(
         Long restaurantId
