+++ src/main/java/com/micorservice/users/application/handler/IUserHandler.java
@@ -1,6 +1,7 @@
 package com.micorservice.users.application.handler;
 
-import com.micorservice.users.application.dto.request.RestaurantIdResponseDto;
+import com.micorservice.users.application.dto.response.PhoneResponseDto;
+import com.micorservice.users.application.dto.response.RestaurantIdResponseDto;
 import com.micorservice.users.application.dto.request.UserRequestDto;
 import com.micorservice.users.application.dto.response.SaveMessageResponse;
 
@@ -11,5 +12,8 @@ public interface IUserHandler {
     void validateUserRole(Long userId, String expectedRole);
 
     RestaurantIdResponseDto getRestaurantByUser(Long employeeId);
+
+    PhoneResponseDto getPhoneNumberByUserId(Long customerId);
+
 
 }
