+++ src/main/java/com/micorservice/users/domain/api/IUserServicePort.java
@@ -9,6 +9,7 @@ public interface IUserServicePort {
     void validateUserRole(Long userId, String expectedRole);
 
     Long getRestaurantByUser(Long employeeId);
+    String getPhoneNumberByUserId(Long customerId);
 
 }
