+++ src/main/java/com/micorservice/users/domain/spi/IUserPersistencePort.java
@@ -23,5 +23,8 @@ public interface IUserPersistencePort {
 
     Long getRestaurantByUser(Long employeeId);
 
+    String getPhoneNumberByUserId(Long customerId);
+
+
 
 }
